refactor(docs): replace barrel files with direct leaf-module imports
