--- apps/docs/src/components/page/article/content/index.ts
@@ -1,1 +1,0 @@
-export * from './ArticleContent'
