--- apps/docs/src/components/page/index.ts
@@ -1,1 +1,0 @@
-export * from './PageBuilder'
