--- apps/docs/src/lib/arcade/index.ts
@@ -1,1 +1,0 @@
-export * from './helpers'
