--- apps/docs/src/lib/codeEditor/index.ts
@@ -1,2 +1,0 @@
-export * from './CodeEditor'
-export * from './types'
