--- apps/docs/src/lib/data/api/index.ts
@@ -1,2 +1,0 @@
-export * from './constants'
-export * from './query'
