--- apps/docs/src/lib/data/index.ts
@@ -1,3 +1,0 @@
-export * from './_global'
-export * from './api'
-export * from './target'
