--- apps/docs/src/lib/nav/index.ts
@@ -1,2 +1,0 @@
-export * from './parseNav'
-export * from './types'
