--- apps/docs/src/studio/schema/article/content/index.ts
@@ -1,4 +1,0 @@
-export * from './blocks'
-export * from './codeExampleField'
-export * from './npmPackageBadgeField'
-export * from './propertyTableField'
